requirements.txt
django-grappelli
django
django-import-export